+++ conda/cli/main_list.py
@@ -11,7 +11,7 @@ from .common import disp_features, stdout_json
 from ..base.constants import DEFAULTS_CHANNEL_NAME, UNKNOWN_CHANNEL
 from ..base.context import context
 from ..common.compat import text_type
-from ..core.linked_data import PrefixData, is_linked, linked
+from ..core.prefix_data import PrefixData, is_linked, linked
 from ..egg_info import get_egg_info
 from ..gateways.disk.test import is_conda_environment
 from ..history import History
