+++ conda/core/prefix_data.py
