+++ conda/install.py
@@ -29,7 +29,7 @@ from .base.constants import PREFIX_PLACEHOLDER
 from .common.compat import on_win, open
 from .gateways.disk.delete import delete_trash, move_path_to_trash, rm_rf
 delete_trash, move_path_to_trash = delete_trash, move_path_to_trash
-from .core.linked_data import is_linked, linked, linked_data  # NOQA
+from .core.prefix_data import is_linked, linked, linked_data  # NOQA
 is_linked, linked, linked_data = is_linked, linked, linked_data
 from .core.package_cache_data import rm_fetched  # NOQA
 rm_fetched = rm_fetched  # lgtm [py/redundant-assignment]
